Rename new implementation files
